add macro for fixed size hash types
